Rename util module to common in new util module
